@@ -15,6 +15,7 @@ namespace nats_client_metrics
 {
     public class Startup
     {
+        readonly string corsAllowAllOrigins = "_corsAllowAllOrigins";
         public Startup(IConfiguration configuration)
         {
             Configuration = configuration;
@@ -25,45 +26,31 @@ namespace nats_client_metrics
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            // ********************
-            // USE CORS
-            // ********************
-            services.AddCors(options =>
+           services.AddCors(options =>
             {
-                options.AddPolicy("AllowAll",
+                options.AddPolicy(name: corsAllowAllOrigins,
                     builder =>
                     {
-                        builder
-                        .AllowAnyOrigin() 
-                        .AllowAnyMethod()
-                        .AllowAnyHeader()
-                        .AllowCredentials();
+                        builder.AllowAnyHeader().AllowAnyMethod().AllowAnyOrigin();
                     });
             });
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2)
-                .AddXmlSerializerFormatters();
+            // map to the application API endpoints
+            services.AddControllers();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
-
-            // ********************
-            // USE CORS
-            // ********************
-            app.UseCors("AllowAll");
+            
             app.UseHttpsRedirection();
-            app.UseMvc();
+            app.UseRouting();
+            // CORS setup has to go here
+            app.UseCors(corsAllowAllOrigins);
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
